+++ openeo/rest/datacube.py
@@ -17,7 +17,18 @@ import typing
 import urllib.parse
 import warnings
 from builtins import staticmethod
-from typing import Any, Callable, Dict, Iterable, List, Optional, Sequence, Tuple, Union
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    Iterable,
+    List,
+    Mapping,
+    Optional,
+    Sequence,
+    Tuple,
+    Union,
+)
 
 import numpy as np
 import requests
@@ -2386,6 +2397,7 @@ class DataCube(_ProcessGraphAbstraction):
         auto_add_save_result: bool = True,
         additional: Optional[dict] = None,
         job_options: Optional[dict] = None,
+        on_response_headers: Optional[Callable[[Mapping], None]] = None,
     ) -> Union[None, bytes]:
         """
         Send the underlying process graph to the backend
@@ -2403,6 +2415,7 @@ class DataCube(_ProcessGraphAbstraction):
         :param additional: (optional) additional (top-level) properties to set in the request body
         :param job_options: (optional) dictionary of job options to pass to the backend
             (under top-level property "job_options")
+        :param on_response_headers: (optional) callback to handle/show the response headers
 
         :return: if ``outputfile`` was not specified:
             a :py:class:`bytes` object containing the raw data.
@@ -2413,6 +2426,9 @@ class DataCube(_ProcessGraphAbstraction):
 
         .. versionchanged:: 0.36.0
             Added arguments ``additional`` and ``job_options``.
+
+        .. versionchanged:: 0.40
+            Added argument ``on_response_headers``.
         """
         # TODO #278 centralize download/create_job/execute_job logic in DataCube, VectorCube, MlModel, ...
         if auto_add_save_result:
@@ -2420,7 +2436,12 @@ class DataCube(_ProcessGraphAbstraction):
         else:
             res = self
         return self._connection.download(
-            res.flat_graph(), outputfile=outputfile, validate=validate, additional=additional, job_options=job_options
+            res.flat_graph(),
+            outputfile=outputfile,
+            validate=validate,
+            additional=additional,
+            job_options=job_options,
+            on_response_headers=on_response_headers,
         )
 
     def validate(self) -> List[dict]: